Added Shortcut and PS1 file for automatic file unblocking
As ExBuddy is recommended to be installed for new users via zip download, files will need to be unblocked once extracted. I have added a powershell script and a shortcut to open powershell and run a script checking the install location and if correct unblocking all files recursively within the Exbuddy folder.